requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
docx2pdf==0.1.7
nltk==3.5
xlrd==1.2.0
numpy==1.19.4
xlwt==1.3.0
docx2txt==0.8
pytesseract==0.3.7
Pillow==8.0.1
pyresparser==1.0.6